requirements.txt
selenium
pytest~=7.3.1
webdriver-manager
Faker ~= 19.1.0
allure-pytest